+++ axum-macros/src/attr_parsing.rs
@@ -82,6 +82,21 @@ where
     Ok(())
 }
 
+pub(crate) fn combine_unary_attribute<K>(a: &mut Option<K>, b: Option<K>) -> syn::Result<()>
+where
+    K: ToTokens,
+{
+    if let Some(kw) = b {
+        if a.is_some() {
+            let kw_name = std::any::type_name::<K>().split("::").last().unwrap();
+            let msg = format!("`{}` specified more than once", kw_name);
+            return Err(syn::Error::new_spanned(kw, msg));
+        }
+        *a = Some(kw);
+    }
+    Ok(())
+}
+
 pub(crate) fn second<T, K>(tuple: (T, K)) -> K {
     tuple.1
 }