+++ axum/src/response/mod.rs
@@ -98,7 +98,7 @@ mod tests {
             }
         }
 
-        Router::<_, Body>::new()
+        Router::<(), Body>::new()
             .route("/", get(impl_trait_ok))
             .route("/", get(impl_trait_err))
             .route("/", get(impl_trait_both))
@@ -208,7 +208,7 @@ mod tests {
             )
         }
 
-        Router::<_, Body>::new()
+        Router::<(), Body>::new()
             .route("/", get(status))
             .route("/", get(status_headermap))
             .route("/", get(status_header_array))