+++ examples/key-value-store/src/main.rs
@@ -26,7 +26,7 @@ use std::{
 use tower::{BoxError, ServiceBuilder};
 use tower_http::{
     auth::RequireAuthorizationLayer, compression::CompressionLayer, limit::RequestBodyLimitLayer,
-    trace::TraceLayer,
+    trace::TraceLayer, ServiceBuilderExt,
 };
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
@@ -43,22 +43,24 @@ async fn main() {
     let shared_state = SharedState::default();
 
     // Build our application by composing routes
-    let app = Router::with_state(Arc::clone(&shared_state))
+    let app = Router::new()
         .route(
             "/:key",
             // Add compression to `kv_get`
             get(kv_get.layer(CompressionLayer::new()))
                 // But don't compress `kv_set`
                 .post_service(
-                    ServiceBuilder::new()
-                        .layer(DefaultBodyLimit::disable())
-                        .layer(RequestBodyLimitLayer::new(1024 * 5_000 /* ~5mb */))
-                        .service(kv_set.with_state(Arc::clone(&shared_state))),
+                    kv_set
+                        .layer((
+                            DefaultBodyLimit::disable(),
+                            RequestBodyLimitLayer::new(1024 * 5_000 /* ~5mb */),
+                        ))
+                        .with_state(Arc::clone(&shared_state)),
                 ),
         )
         .route("/keys", get(list_keys))
         // Nest our admin routes under `/admin`
-        .nest("/admin", admin_routes(shared_state))
+        .nest("/admin", admin_routes())
         // Add middleware to all routes
         .layer(
             ServiceBuilder::new()
@@ -67,9 +69,9 @@ async fn main() {
                 .load_shed()
                 .concurrency_limit(1024)
                 .timeout(Duration::from_secs(10))
-                .layer(TraceLayer::new_for_http())
-                .into_inner(),
-        );
+                .layer(TraceLayer::new_for_http()),
+        )
+        .with_state(Arc::clone(&shared_state));
 
     // Run our app with hyper
     let addr = SocketAddr::from(([127, 0, 0, 1], 3000));
@@ -113,7 +115,7 @@ async fn list_keys(State(state): State<SharedState>) -> String {
         .join("\n")
 }
 
-fn admin_routes(state: SharedState) -> Router<SharedState> {
+fn admin_routes() -> Router<SharedState> {
     async fn delete_all_keys(State(state): State<SharedState>) {
         state.write().unwrap().db.clear();
     }
@@ -122,7 +124,7 @@ fn admin_routes(state: SharedState) -> Router<SharedState> {
         state.write().unwrap().db.remove(&key);
     }
 
-    Router::with_state(state)
+    Router::new()
         .route("/keys", delete(delete_all_keys))
         .route("/key/:key", delete(remove_key))
         // Require bearer auth for all admin routes